apps/requirements.txt
Django==2.1
gunicorn==19.9.0
pytz==2018.9
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
defusedxml==0.5.0
django-allauth==0.36.0
django-auth-adfs==0.2.1
django-constance==2.3.1
django-cors-headers==2.2.0
django-extensions==2.1.3
django-filter==2.0.0
django-filters==0.2.1
django-picklefield==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-role-filters==0.2.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
llvmlite==0.25.0
MarkupSafe==1.0
numba==0.40.0
numpy==1.15.1
oauthlib==2.1.0
openapi-codec==1.3.2
pandas==0.23.4
pycparser==2.19
pydash==4.7.3
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pyOpenSSL==18.0.0
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23

